Add a grid renderer
This class is responsible to output a given grid to the terminal.
It simply iterates over each row and column and display the value at
this position in the grid.
It properly adds a line after the first row and a pipe after the first
column.
Each value is centered based on the biggest value of the grid, with a
margin of 1 space on each side.